requirements.txt
pybuilder
pycryptodome
fastapi
uvicorn